More cleanup and add Macros
There is a lot of changes here but most of it is because more is
being generated now.

This adds a macro feature that will expand out the validation lines.
This makes the yaml a little more compact and readable.  It also
removes the need to specify the noop for every key.

This also fixes a number of minor errors from before.

A new schema was added for the macro file format and the validation
script was modified to take a schema and list of files so it is more
general.